+++ CHANGELOG.md
@@ -55,6 +55,8 @@ with only minimum code changes required. See [the MD documentation](docs/bring_y
   (`ScalarLoss.CustomClassification` and `CustomRegression`), prediction targets (`ScalarModelBase.target_names`),
   and reporting (`ModelConfigBase.generate_custom_report()`) in scalar configs, providing more flexibility for defining
   model configs with custom behaviour while leveraging the existing InnerEye workflows.
+- ([#445](https://github.com/microsoft/InnerEye-DeepLearning/pull/445)) Adding test coverage for the `HelloContainer`
+  model with multiple GPUs
 
 ### Changed
 
@@ -83,6 +85,8 @@ with only minimum code changes required. See [the MD documentation](docs/bring_y
   set, display an error message and terminate the run.
 - ([#437](https://github.com/microsoft/InnerEye-DeepLearning/pull/437)) Fixed multi-node DDP bug in PL v1.2.8. Re-add
   end-to-end test for multi-node.
+- ([#445](https://github.com/microsoft/InnerEye-DeepLearning/pull/445)) Fixed a bug when running inference for
+ container models on machines with >1 GPU
 
 ### Removed
 - ([#439](https://github.com/microsoft/InnerEye-DeepLearning/pull/439)) Deprecated `start_epoch` config argument.